requirements.txt
transformers
torch
colbert-ai